[1/N] Quantization refactor: split AutoRound into package layout
Move AutoRound from a single module into quantization/auto_round/.
No behavior change.
