+++ python/sglang/srt/layers/quantization/auto_round/__init__.py
@@ -1,0 +1,5 @@
+# SPDX-License-Identifier: Apache-2.0
+
+from .auto_round import AutoRoundConfig
+
+__all__ = ["AutoRoundConfig"]
